Switching default screen to login
