+++ app/src/main/java/com/parcelpaladin/d30f2n/parcelpaladinfirebase/MainActivity.java
@@ -9,7 +9,6 @@ import android.text.TextUtils;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.ProgressBar;
 import android.widget.TextView;
 import android.widget.Toast;
 
@@ -17,21 +16,13 @@ import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
-
-import org.w3c.dom.Text;
 
 public class MainActivity extends AppCompatActivity implements View.OnClickListener{
 
-    private Button buttonRegister;
-    private EditText editTextName;
+    private Button buttonSignIn;
     private EditText editTextEmail;
     private EditText editTextPassword;
-    private TextView textViewSignIn;
-
-    private DatabaseReference mDatabase;
+    private TextView textViewSignUp;
 
     private ProgressDialog progressDialog;
 
@@ -46,28 +37,26 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
 
         if(firebaseAuth.getCurrentUser() != null)
         {
-            //user already logged in
+            //profile activity here
             finish();
             startActivity(new Intent(getApplicationContext(), ProfileActivity.class));
         }
 
-
         progressDialog = new ProgressDialog(this);
-
-        buttonRegister = (Button) findViewById(R.id.buttonRegister);
         editTextEmail = (EditText) findViewById(R.id.editTextEmail);
         editTextPassword = (EditText) findViewById(R.id.editTextPassword);
-        editTextName = (EditText) findViewById(R.id.editTextName);
+        buttonSignIn = (Button) findViewById(R.id.buttonSignin);
+        textViewSignUp = (TextView) findViewById(R.id.textViewSignUp);
 
-        textViewSignIn = (TextView) findViewById(R.id.textViewSignIn);
 
-        buttonRegister.setOnClickListener(this);
-        textViewSignIn.setOnClickListener(this);
+
+        buttonSignIn.setOnClickListener(this);
+        textViewSignUp.setOnClickListener(this);
+
     }
 
-    private void registerUser()
+    private void userLogin()
     {
-        final String name = editTextName.getText().toString().trim();
         String email = editTextEmail.getText().toString().trim();
         String password = editTextPassword.getText().toString().trim();
 
@@ -87,27 +76,17 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         progressDialog.setMessage("Registering...");
         progressDialog.show();
 
-        firebaseAuth.createUserWithEmailAndPassword(email, password)
+        firebaseAuth.signInWithEmailAndPassword(email, password)
                 .addOnCompleteListener(this, new OnCompleteListener<AuthResult>() {
                     @Override
                     public void onComplete(@NonNull Task<AuthResult> task) {
+                        progressDialog.dismiss();
+
                         if(task.isSuccessful())
                         {
-                            Toast.makeText(MainActivity.this, "Registration Successful", Toast.LENGTH_SHORT).show();
-                            FirebaseUser user = firebaseAuth.getCurrentUser();
-                            mDatabase = FirebaseDatabase.getInstance().getReference("users");
-                            mDatabase.child(user.getUid());
-                            UserInformation userInformation = new UserInformation(name);
-                            mDatabase = FirebaseDatabase.getInstance().getReference("users/"+user.getUid());
-                            mDatabase.child("Name").setValue(name);
-                            mDatabase.child("Email").setValue(user.getEmail());
-                            mDatabase.child("trackingNumbers");
+                            //start profile activity
                             finish();
                             startActivity(new Intent(getApplicationContext(), ProfileActivity.class));
-                        }
-                        else
-                        {
-                            Toast.makeText(MainActivity.this, "Registration Failed", Toast.LENGTH_SHORT).show();
                         }
                     }
                 });
@@ -115,16 +94,15 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
     }
 
     @Override
-    public void onClick(View view) {
-        if(view==buttonRegister)
+    public void onClick(View v) {
+        if(v==buttonSignIn)
         {
-            registerUser();
+            userLogin();
         }
-        if(view==textViewSignIn)
+        if(v == textViewSignUp)
         {
-            //Will open login activity here
-            startActivity(new Intent(this, LoginActivity.class));
+            finish();
+            startActivity(new Intent(this, RegistrationActivity.class));
         }
-
     }
 }
