+++ app/src/main/res/layout/activity_registration.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.parcelpaladin.d30f2n.parcelpaladinfirebase.LoginActivity">
+    tools:context="com.parcelpaladin.d30f2n.parcelpaladinfirebase.RegistrationActivity">
 
     <LinearLayout
         android:layout_centerHorizontal="true"
@@ -17,12 +17,21 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_gravity="center"
-            android:text="User Login"
-            android:id="@+id/userLogin"
+            android:text="User Registration"
+            android:id="@+id/userReg"
             android:textAppearance="@style/TextAppearance.AppCompat.Display2" />
 
         <EditText
 
+            android:layout_margin="15dp"
+            android:inputType="textPersonName"
+            android:hint="Enter your name"
+            android:id="@+id/editTextName"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content" />
+
+
+        <EditText
             android:layout_margin="15dp"
             android:inputType="textEmailAddress"
             android:hint="Enter your email"
@@ -33,21 +42,21 @@
         <EditText
             android:layout_margin="15dp"
             android:inputType="textPassword"
-            android:hint="Enter your password"
+            android:hint="Create a password"
             android:id="@+id/editTextPassword"
             android:layout_width="match_parent"
             android:layout_height="wrap_content" />
 
         <Button
-            android:id="@+id/buttonSignin"
+            android:id="@+id/buttonRegister"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:text="Sign in" />
+            android:text="Register" />
 
         <TextView
             android:textAlignment="center"
-            android:text="Haven't registered yet? Sign up here"
-            android:id="@+id/textViewSignUp"
+            android:text="Already registered? Sign in here"
+            android:id="@+id/textViewSignIn"
             android:layout_width="match_parent"
             android:layout_height="wrap_content" />
